Add a line to WPXA.ahk to #include WinGetPosEx
Trying to run WindowPadX was failing (at least in AHK v1.1) with the error
that it was trying to call the nonexistent function WinGetPosEx.

This was due to a missing #include line to the file WinGetPosEx.ahk. This
has now been added, and the script can run and compile again.